Resolves Orientation Issues
[ We live in complicated fluctuating times, that have complicated problems, which require complicated solutions, so please bare with me here, thanks.]
Okay so now I finally got this one sorted out a little better. For a number of years it seems this has been dysfunctional and some of that has been related to spec and browser stability, which seems likely to remain ongoing.
My recent path of discovery uncovering this in part https://github.com/mrdoob/three.js/issues/12789
and initial attempts to resolve this for the record are here
https://github.com/mrdoob/three.js/pull/12798
https://github.com/mrdoob/three.js/pull/12850
That's the one @WestLangley requested to be resubmitted, which this does.

Here's also a small list of seemingly related issues that have come up on this in the past and appear when searching on the mysterious hard to reproduce problem for cross-reference, (that seems to occur for all people platforms languages etc. )
https://github.com/mrdoob/three.js/issues/7502
https://github.com/mrdoob/three.js/issues/10552
https://stackoverflow.com/questions/36314415/add-offset-to-deviceorientationcontrols-in-three-js

The bad user experience is that the normal 'front and center' of the browser experience or initial camera positioning is random and not predictable or stable on startup.

The problem is actually three-fold:
First (and not as obvious possibly 'weakest [slowest legacy device] link' issue) when a scene is starting-up/initializing, the device orientation (or angles of the rotational positioning of the device) are not initially stable, so this PR adds a little delay on updating the computing the initial alphaOffset (waiting for accumulated stability in the readings),
[It is not related to 'absolute' magnetic compass orientation as that is not desired and added more recently to help with these issues. What we want is for the user to get the correct initial camera orientation intended you get with the browser "face-forward" regardless of their actual physical orientation.
[ newer 'absolute' spec reference https://developer.mozilla.org/en-US/docs/Web/API/Detecting_device_orientation ]

The second issue is that alpha offset is greatly affected by gamma or tilt (so z-rotation or camera direction rotations) {being tangled into it's Quaternion aka matrix of course they are correlated let me explain further.}
Overall when the device is well balanced the alpha/heading fluctuates wildly (spinning on it's well centered Y-axis). If this is set as the alphaOffset on start up it can be essentially random, without skipping those perfectly balanced beta values (aka y-axis, vertical with gravity) that in the Matrix (Quaternion) they are associated with cause random alpha values it will cause this kind of random heading problem. Note that because this is extremely difficult or rare over a few samples it is never perfectly vertical in practice and so not only fine to skip those but essential.
The third has to do with  Portrait vs Landscape  screen orientation. On my Galaxy Note 8 (Android 7) but not the Note 3 (Android 5) [which also doesn't support WebVR] there is a slight offset still needed in landscape mode compared with portrait mode. Maybe that's a bug to report as well? Alas it seems there could be a need for device & OS version specific offsets?! [I'm not sure how these can be calibrated and validated on the browser/spec hardware level, but feel it's lacking and needed.]
The gist of this is I added a deviceLandscapeAlphaOffset that can be set via 'your' code externally, which I suppose is kind of like alphaOffset was intended but we want this calculated correctly for us on startup, I'm thinking 100% of the time and so setting alphaOffset manually at start would require setting offChks above 7 which is the number of times I randomly choose to hopefully achieve stable readings. (so that too is available on scope).
I'm assuming the screenOrientation values sent are based on averages, so I didn't do an averaging in this code myself as that would never achieve the correct setting, while waiting too long is also not desired if it is correct out of the gate, but it needs to be set relatively quickly and not change henceforth.
[I know it's like I say at the top not an pretty fix to a complicated problem.]
In the future maybe the offChks for stability is flaky and becomes obsolete. Also one hopes there's a calibration process or the spec maintainers provide a better way to provide consistency and stability across devices. Then deviceLandscapeAlphaOffset would seem unneeded.

Final words about other changes to this file:
I've left 'var's as that was requested (in my previous attempt at solving this disaster) but I really believe strongly in the none verbatim words of Crockford "Sorry 'var' was bad we fixed it with 'let', but it breaks things so 'var' had to stay broken, but please don't use it moving forward because it's bad." New performance tests also show in the latest iterations of V8/Chrome 'let' is also pico seconds better then 'var'. I think this is a perfect place to start allowing 'let' to become standard practice.

Also I liked the use of 'scope' over 'this' so I stomped all 'this' with scope throughout, from what I understand it's more reliable/better too.
To further abolish the usage of the "var vs let" issue I use the 'scope' variable to set/export the functions like 'update' 'connect' etc.
Variables used twice or more get a local copy (typically in my books & code), in the same token things used once even for clarity don't get stored locally inside a function call (see changes to update function).

I left your setObjectQuaternion function scope wrapper with private 'var's for zee, euler & Quaternions. [discussion points omitted here]
Reduced some redundancy in naming convention lengths and empty line spaces to increase legibility being only a pages worth.

The previous version had compact 'if' statements in the 'update' function and I'm pretty sure are not needed thanks to the low probability deviceOrientation is only partially set that would be a browser spec bug and simply redundant to check per variable. Anyway works fine without it in my tests (much nicer/cleaner declaration setting checking solution there). Maybe it was a legacy issue, but it seemed cumbersome, undoubtedly slower and totally redundant at this point. Alas I'm the first to say initially set all argument's to 'defaults' mainly so all functions never fail and even expend that notion to external caching for things like zee in setObjectQuaternion if that shows performance enhancements, especially without the other fundamental design improvements therein, without stack loading scope wrapped argument values etc. [despite it flying in the face of functional programming logic, because we now those values are not changing {further it allows resetting of cached values passable as arguments...being JS is single threaded all this seemingly Socratic reverse thinking is fine. {Alas there is no "advanced class" on that stuff, so I leave it to the reader to contemplate}].

Anyway this works better in many ways for me and hopefully others. Although it's complicated and slightly ugly it does work to get a face-forward (browser consistent) user experience when a scene loads with orientation controls. I'm not saying it's perfect, just better for me and felt compelled to share.

Sorry I missed 89r on this one.
Congratulations on that and I wish for many Happy Thankful Holidays to all supporting "THREE js ~ The future of computing today."